Moved all code out of the root com.stripe namespace. Refactored some package names.
