+++ src/main/java/com/stripe/sdk/exception/StripeBadRequestException.java
@@ -1,4 +1,4 @@
-package com.stripe.exceptions;
+package com.stripe.sdk.exception;
 
 public class StripeBadRequestException extends StripeException {
 	public StripeBadRequestException(){
