+++ src/main/java/com/stripe/sdk/exception/StripeNotFoundException.java
@@ -1,4 +1,4 @@
-package com.stripe.exceptions;
+package com.stripe.sdk.exception;
 
 public class StripeNotFoundException extends StripeException {
 	public StripeNotFoundException(){
