+++ src/main/java/com/stripe/sdk/model/Charge.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class Charge {
 	protected boolean attempted;
