+++ src/main/java/com/stripe/sdk/model/Customer.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class Customer {
 	protected String description;
