+++ src/main/java/com/stripe/sdk/model/Invoice.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class Invoice {
 	protected long created, subtotal, total;
