+++ src/main/java/com/stripe/sdk/model/InvoiceItem.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class InvoiceItem {
 	protected long date, amount;
