+++ src/main/java/com/stripe/sdk/model/Period.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class Period {
 	protected long start, end;
