+++ src/main/java/com/stripe/sdk/model/StripeError.java
@@ -1,4 +1,4 @@
-package com.stripe.datamodel;
+package com.stripe.sdk.model;
 
 public class StripeError {
 	public String code;
